+++ Engine/Group.cpp
@@ -31,7 +31,7 @@ namespace Engine {
         }
         controls.clear();
     }
-    void Group::Update(float deltaTime) {
+    void Group::Update(double deltaTime) {
         for (auto it = objects.begin(); it != objects.end();) {
             auto preIt = it++;
             if (preIt->second->Visible)
